requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.0.3
flask-cors==6.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
gunicorn==21.2.0
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0